multi-modal content for chat completions
The OpenAI API supports two kinds of Content for chat completion: Text
and Image URLs (including base64 encoded data URLs). The approach this
change takes is to support only multi-modal (the super set of text
generation) rather than attempt to serialization tricks to handle both.
This is breaking and probably not generally as supportable.